+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unpublished Changes
+## 0.0.16
 
 - Support dart 2 runtimes
 - Better error message for when attribute selectors have an operator but no